+++ lib/modules/core/core_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: core
-// Updated at: 2026-04-09 14:08:46
-// Commit: test(core): implement widget integration tests for core
+// Updated at: 2026-04-10 09:59:05
+// Commit: chore(core): update dependencies and cleanup unused imports in core
 
 class coreComponentState {
-  final String commitRef = '225';
-  final String lastMessage = 'test(core): implement widget integration tests for core';
-  final DateTime timestamp = DateTime.parse('2026-04-09T14:08:46+05:00');
+  final String commitRef = '263';
+  final String lastMessage = 'chore(core): update dependencies and cleanup unused imports in core';
+  final DateTime timestamp = DateTime.parse('2026-04-10T09:59:05+05:00');
 }